requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Azure services
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-openai==1.0.0

# Document processing and embeddings
openai==1.3.7
tiktoken==0.5.2
PyPDF2==3.0.1
python-multipart==0.0.6

# Data processing
pandas==2.1.4
numpy==1.24.3
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0